Added evaluation.RunGames as main point of entry
